requirements.txt
Flask==2.1.2
gunicorn==19.9.0
flask_cors==3.0.10
Jinja2==3.1.2

beautifulsoup4>=4.12.2
colorama==0.4.6
distro==1.8.0
openai==0.27.2
playsound==1.2.2
python-dotenv
pyyaml==6.0
readability-lxml
requests
gTTS==2.3.1
docker
duckduckgo-search
google-api-python-client #(https://developers.google.com/custom-search/v1/overview)
pinecone-client==2.2.1

jsonlines==2.0.0
langchain
pdfplumber

supabase
